pkg/cryptoauthlib: fix native build
Give priority to RIOT_APPLICATION macro over __linux__ and __WIN32__
ones.

+++ sys/auto_init/security/auto_init_atca.c
@@ -16,6 +16,8 @@
  * @}
  */
 
+#include <inttypes.h>
+
 #include "log.h"
 #include "atca.h"
 #include "atca_params.h"
@@ -66,7 +68,7 @@ void auto_init_atca(void)
 #else
 void auto_init_atca(void)
 {
-    DEBUG("[auto_init_atca] Number of secure elements: %d\n", ATCA_NUMOF);
+    DEBUG("[auto_init_atca] Number of secure elements: %"PRIu32"\n", (uint32_t)ATCA_NUMOF);
     for (unsigned i = 0; i < ATCA_NUMOF; i++) {
         int status = initATCADevice((ATCAIfaceCfg *)&atca_params[i], (ATCADevice)&atca_devs[i]);
         if (status != ATCA_SUCCESS) {
